youtube/requirements.txt
Django==4.2.2
django-import-export==3.2.0
django-jazzmin==2.6.0
django-taggit==4.0.0

flake8==6.1.0
isort==5.12.0

Pillow==9.5.0
